+++ runtime-light/stdlib/stdlib.cmake
@@ -2,11 +2,10 @@ prepend(
   RUNTIME_STDLIB_SRC
   stdlib/
   component/component-api.cpp
-  exit/exit.cpp
-  fork/fork-api.cpp
+  exit/exit-functions.cpp
   fork/fork-context.cpp
   output/output-buffer.cpp
-  output/print.cpp
+  output/print-functions.cpp
   rpc/rpc-api.cpp
   rpc/rpc-context.cpp
   rpc/rpc-extra-headers.cpp
